# blas/gonum: improve dgemmParallel performance