--- a/blas/gonum/gonum.go
+++ b/blas/gonum/gonum.go
@@ -19,13 +19,7 @@ type Implementation struct{}
 const (
 	blockSize   = 64 // b x b matrix
 	minParBlock = 4  // minimum number of blocks needed to go parallel
-	buffMul     = 4  // how big is the buffer relative to the number of workers
 )
-
-// subMul is a common type shared by [SD]gemm.
-type subMul struct {
-	i, j int // index of block
-}
 
 func max(a, b int) int {
 	if a > b {