@kortschak 2020-03-13
Egon, can you run `go generate` in the blas/gonum package, please. This will generate the float32 version of the code.